huhsame_backend/requirements.txt
fastapi
uvicorn
langchain
langchain-community
langchain-openai
openai
python-dotenv